requirements.txt
aiofiles==0.8.0
asyncio==3.4.3
beautifulsoup4==4.10.0
cryptography==3.4.7
gitpython==3.1.24
github3.py==2.0.0
jwt==1.3.1
markdown2==2.4.0
pdoc3==0.10.0
psutil==5.8.0
PyGithub==1.55
pytest==6.2.5
pytest-asyncio==0.15.1
python-dateutil==2.8.2
requests==2.26.0
astor==0.8.1
redis
python-magic 
Pillow 
pdf2image





